Network.AWS.CodePipeline.CreateCustomActionType

Description

Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.

Creating a Request

createCustomActionType

Arguments

:: ActionCategory
ccatCategory

-> Text
ccatProvider

-> Text
ccatVersion

-> ArtifactDetails
ccatInputArtifactDetails

-> ArtifactDetails
ccatOutputArtifactDetails

-> CreateCustomActionType

Creates a value of CreateCustomActionType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccatSettings - Returns information about the settings for an action type.

ccatConfigurationProperties - The configuration properties for the custom action.

ccatCategory - The category of the custom action, such as a build action or a test action.

ccatProvider - The provider of the service used in the custom action, such as AWS CodeDeploy.

ccatVersion - The version identifier of the custom action.

ccatInputArtifactDetails - The details of the input artifact for the action, such as its commit ID.

ccatOutputArtifactDetails - The details of the output artifact of the action, such as its commit ID.

data CreateCustomActionType

Represents the input of a create custom action operation.

See: createCustomActionType smart constructor.

Request Lenses

ccatSettings :: Lens' CreateCustomActionType (Maybe ActionTypeSettings)

Returns information about the settings for an action type.

ccatConfigurationProperties :: Lens' CreateCustomActionType [ActionConfigurationProperty]

The configuration properties for the custom action.

ccatCategory :: Lens' CreateCustomActionType ActionCategory

The category of the custom action, such as a build action or a test action.

ccatProvider :: Lens' CreateCustomActionType Text

The provider of the service used in the custom action, such as AWS CodeDeploy.

ccatVersion :: Lens' CreateCustomActionType Text

The version identifier of the custom action.

ccatInputArtifactDetails :: Lens' CreateCustomActionType ArtifactDetails

The details of the input artifact for the action, such as its commit ID.

ccatOutputArtifactDetails :: Lens' CreateCustomActionType ArtifactDetails

The details of the output artifact of the action, such as its commit ID.

Destructuring the Response

createCustomActionTypeResponse

Arguments

:: Int
ccatrsResponseStatus

-> ActionType
ccatrsActionType

-> CreateCustomActionTypeResponse

Creates a value of CreateCustomActionTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccatrsResponseStatus - -- | The response status code.

ccatrsActionType - Returns information about the details of an action type.

data CreateCustomActionTypeResponse

Represents the output of a create custom action operation.

See: createCustomActionTypeResponse smart constructor.

Response Lenses

ccatrsResponseStatus :: Lens' CreateCustomActionTypeResponse Int

- | The response status code.

ccatrsActionType :: Lens' CreateCustomActionTypeResponse ActionType

Returns information about the details of an action type.