Network.AWS.CodePipeline.DisableStageTransition

Description

Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.

Creating a Request

disableStageTransition

Arguments

:: Text
dstPipelineName

-> Text
dstStageName

-> StageTransitionType
dstTransitionType

-> Text
dstReason

-> DisableStageTransition

Creates a value of DisableStageTransition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dstPipelineName - The name of the pipeline in which you want to disable the flow of artifacts from one stage to another.

dstStageName - The name of the stage where you want to disable the inbound or outbound transition of artifacts.

dstTransitionType - Specifies whether artifacts will be prevented from transitioning into the stage and being processed by the actions in that stage (inbound), or prevented from transitioning from the stage after they have been processed by the actions in that stage (outbound).

dstReason - The reason given to the user why a stage is disabled, such as waiting for manual approval or manual tests. This message is displayed in the pipeline console UI.

data DisableStageTransition

Represents the input of a disable stage transition input action.

See: disableStageTransition smart constructor.

Request Lenses

dstPipelineName :: Lens' DisableStageTransition Text

The name of the pipeline in which you want to disable the flow of artifacts from one stage to another.

dstStageName :: Lens' DisableStageTransition Text

The name of the stage where you want to disable the inbound or outbound transition of artifacts.

dstTransitionType :: Lens' DisableStageTransition StageTransitionType

Specifies whether artifacts will be prevented from transitioning into the stage and being processed by the actions in that stage (inbound), or prevented from transitioning from the stage after they have been processed by the actions in that stage (outbound).

dstReason :: Lens' DisableStageTransition Text

The reason given to the user why a stage is disabled, such as waiting for manual approval or manual tests. This message is displayed in the pipeline console UI.

Destructuring the Response

disableStageTransitionResponse :: DisableStageTransitionResponse

Creates a value of DisableStageTransitionResponse with the minimum fields required to make a request.

data DisableStageTransitionResponse

See: disableStageTransitionResponse smart constructor.