Network.AWS.CodePipeline.GetJobDetails

Description

Returns information about a job. Only used for custom actions.

Important: When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Creating a Request

getJobDetails

Arguments

:: Text
gjdJobId

-> GetJobDetails

Creates a value of GetJobDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gjdJobId - The unique system-generated ID for the job.

data GetJobDetails

Represents the input of a get job details action.

See: getJobDetails smart constructor.

Request Lenses

gjdJobId :: Lens' GetJobDetails Text

The unique system-generated ID for the job.

Destructuring the Response

getJobDetailsResponse

Arguments

:: Int
gjdrsResponseStatus

-> GetJobDetailsResponse

Creates a value of GetJobDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gjdrsJobDetails - The details of the job.

gjdrsResponseStatus - -- | The response status code.

data GetJobDetailsResponse

Represents the output of a get job details action.

See: getJobDetailsResponse smart constructor.

Response Lenses

gjdrsJobDetails :: Lens' GetJobDetailsResponse (Maybe JobDetails)

The details of the job.

gjdrsResponseStatus :: Lens' GetJobDetailsResponse Int

- | The response status code.