Network.AWS.CodePipeline.GetPipelineState

Description

Returns information about the state of a pipeline, including the stages and actions.

Creating a Request

getPipelineState

Arguments

:: Text
gpsName

-> GetPipelineState

Creates a value of GetPipelineState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpsName - The name of the pipeline about which you want to get information.

data GetPipelineState

Represents the input of a get pipeline state action.

See: getPipelineState smart constructor.

Request Lenses

gpsName :: Lens' GetPipelineState Text

The name of the pipeline about which you want to get information.

Destructuring the Response

getPipelineStateResponse

Arguments

:: Int
gpsrsResponseStatus

-> GetPipelineStateResponse

Creates a value of GetPipelineStateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpsrsPipelineName - The name of the pipeline for which you want to get the state.

gpsrsCreated - The date and time the pipeline was created, in timestamp format.

gpsrsStageStates - A list of the pipeline stage output information, including stage name, state, most recent run details, whether the stage is disabled, and other data.

gpsrsPipelineVersion - The version number of the pipeline.

gpsrsUpdated - The date and time the pipeline was last updated, in timestamp format.

gpsrsResponseStatus - -- | The response status code.

data GetPipelineStateResponse

Represents the output of a get pipeline state action.

See: getPipelineStateResponse smart constructor.

Response Lenses

gpsrsPipelineName :: Lens' GetPipelineStateResponse (Maybe Text)

The name of the pipeline for which you want to get the state.

gpsrsCreated :: Lens' GetPipelineStateResponse (Maybe UTCTime)

The date and time the pipeline was created, in timestamp format.

gpsrsStageStates :: Lens' GetPipelineStateResponse [StageState]

A list of the pipeline stage output information, including stage name, state, most recent run details, whether the stage is disabled, and other data.

gpsrsPipelineVersion :: Lens' GetPipelineStateResponse (Maybe Natural)

The version number of the pipeline.

gpsrsUpdated :: Lens' GetPipelineStateResponse (Maybe UTCTime)

The date and time the pipeline was last updated, in timestamp format.

gpsrsResponseStatus :: Lens' GetPipelineStateResponse Int

- | The response status code.