Network.AWS.CodePipeline.GetThirdPartyJobDetails

Description

Requests the details of a job for a third party action. Only used for partner actions.

Important: When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Creating a Request

getThirdPartyJobDetails

Arguments

:: Text
gtpjdJobId

-> Text
gtpjdClientToken

-> GetThirdPartyJobDetails

Creates a value of GetThirdPartyJobDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtpjdJobId - The unique system-generated ID used for identifying the job.

gtpjdClientToken - The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

data GetThirdPartyJobDetails

Represents the input of a get third party job details action.

See: getThirdPartyJobDetails smart constructor.

Request Lenses

gtpjdJobId :: Lens' GetThirdPartyJobDetails Text

The unique system-generated ID used for identifying the job.

gtpjdClientToken :: Lens' GetThirdPartyJobDetails Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Destructuring the Response

getThirdPartyJobDetailsResponse

Arguments

:: Int
gtpjdrsResponseStatus

-> GetThirdPartyJobDetailsResponse

Creates a value of GetThirdPartyJobDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtpjdrsJobDetails - The details of the job, including any protected values defined for the job.

gtpjdrsResponseStatus - -- | The response status code.

data GetThirdPartyJobDetailsResponse

Represents the output of a get third party job details action.

See: getThirdPartyJobDetailsResponse smart constructor.

Response Lenses

gtpjdrsJobDetails :: Lens' GetThirdPartyJobDetailsResponse (Maybe ThirdPartyJobDetails)

The details of the job, including any protected values defined for the job.

gtpjdrsResponseStatus :: Lens' GetThirdPartyJobDetailsResponse Int

- | The response status code.