Network.AWS.CodePipeline.ListPipelineExecutions

Description

Gets a summary of the most recent executions for a pipeline.

Creating a Request

listPipelineExecutions

Arguments

:: Text
lpePipelineName

-> ListPipelineExecutions

Creates a value of ListPipelineExecutions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpeNextToken - The token that was returned from the previous list pipeline executions call, which can be used to return the next set of pipeline executions in the list.

lpeMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. The available pipeline execution history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

lpePipelineName - The name of the pipeline for which you want to get execution summary information.

data ListPipelineExecutions

Represents the input of a list pipeline executions action.

See: listPipelineExecutions smart constructor.

Request Lenses

lpeNextToken :: Lens' ListPipelineExecutions (Maybe Text)

The token that was returned from the previous list pipeline executions call, which can be used to return the next set of pipeline executions in the list.

lpeMaxResults :: Lens' ListPipelineExecutions (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. The available pipeline execution history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

lpePipelineName :: Lens' ListPipelineExecutions Text

The name of the pipeline for which you want to get execution summary information.

Destructuring the Response

listPipelineExecutionsResponse

Arguments

:: Int
lpersResponseStatus

-> ListPipelineExecutionsResponse

Creates a value of ListPipelineExecutionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpersNextToken - A token that can be used in the next list pipeline executions call to return the next set of pipeline executions. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

lpersPipelineExecutionSummaries - A list of executions in the history of a pipeline.

lpersResponseStatus - -- | The response status code.

data ListPipelineExecutionsResponse

Represents the output of a list pipeline executions action.

See: listPipelineExecutionsResponse smart constructor.

Response Lenses

lpersNextToken :: Lens' ListPipelineExecutionsResponse (Maybe Text)

A token that can be used in the next list pipeline executions call to return the next set of pipeline executions. To view all items in the list, continue to call this operation with each subsequent token until no more nextToken values are returned.

lpersPipelineExecutionSummaries :: Lens' ListPipelineExecutionsResponse [PipelineExecutionSummary]

A list of executions in the history of a pipeline.

lpersResponseStatus :: Lens' ListPipelineExecutionsResponse Int

- | The response status code.