Network.AWS.CodePipeline.PollForJobs

Description

Returns information about any jobs for AWS CodePipeline to act upon.

Important: When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Creating a Request

pollForJobs

Arguments

:: ActionTypeId
pfjActionTypeId

-> PollForJobs

Creates a value of PollForJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pfjMaxBatchSize - The maximum number of jobs to return in a poll for jobs call.

pfjQueryParam - A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value will be returned.

pfjActionTypeId - Represents information about an action type.

data PollForJobs

Represents the input of a poll for jobs action.

See: pollForJobs smart constructor.

Request Lenses

pfjMaxBatchSize :: Lens' PollForJobs (Maybe Natural)

The maximum number of jobs to return in a poll for jobs call.

pfjQueryParam :: Lens' PollForJobs (HashMap Text Text)

A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value will be returned.

pfjActionTypeId :: Lens' PollForJobs ActionTypeId

Represents information about an action type.

Destructuring the Response

pollForJobsResponse

Arguments

:: Int
pfjrsResponseStatus

-> PollForJobsResponse

Creates a value of PollForJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pfjrsJobs - Information about the jobs to take action on.

pfjrsResponseStatus - -- | The response status code.

data PollForJobsResponse

Represents the output of a poll for jobs action.

See: pollForJobsResponse smart constructor.

Response Lenses

pfjrsJobs :: Lens' PollForJobsResponse [Job]

Information about the jobs to take action on.

pfjrsResponseStatus :: Lens' PollForJobsResponse Int

- | The response status code.