Network.AWS.CodePipeline.PollForThirdPartyJobs

Description

Determines whether there are any third party jobs for a job worker to act on. Only used for partner actions.

Important: When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts.

Creating a Request

pollForThirdPartyJobs

Arguments

:: ActionTypeId
pftpjActionTypeId

-> PollForThirdPartyJobs

Creates a value of PollForThirdPartyJobs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pftpjMaxBatchSize - The maximum number of jobs to return in a poll for jobs call.

pftpjActionTypeId - Represents information about an action type.

data PollForThirdPartyJobs

Represents the input of a poll for third party jobs action.

See: pollForThirdPartyJobs smart constructor.

Request Lenses

pftpjMaxBatchSize :: Lens' PollForThirdPartyJobs (Maybe Natural)

The maximum number of jobs to return in a poll for jobs call.

pftpjActionTypeId :: Lens' PollForThirdPartyJobs ActionTypeId

Represents information about an action type.

Destructuring the Response

pollForThirdPartyJobsResponse

Arguments

:: Int
pftpjrsResponseStatus

-> PollForThirdPartyJobsResponse

Creates a value of PollForThirdPartyJobsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pftpjrsJobs - Information about the jobs to take action on.

pftpjrsResponseStatus - -- | The response status code.

data PollForThirdPartyJobsResponse

Represents the output of a poll for third party jobs action.

See: pollForThirdPartyJobsResponse smart constructor.

Response Lenses

pftpjrsJobs :: Lens' PollForThirdPartyJobsResponse [ThirdPartyJob]

Information about the jobs to take action on.

pftpjrsResponseStatus :: Lens' PollForThirdPartyJobsResponse Int

- | The response status code.