Network.AWS.CodePipeline.PutActionRevision

Description

Provides information to AWS CodePipeline about new revisions to a source.

Creating a Request

putActionRevision

Arguments

:: Text
pPipelineName

-> Text
pStageName

-> Text
pActionName

-> ActionRevision
pActionRevision

-> PutActionRevision

Creates a value of PutActionRevision with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pPipelineName - The name of the pipeline that will start processing the revision to the source.

pStageName - The name of the stage that contains the action that will act upon the revision.

pActionName - The name of the action that will process the revision.

pActionRevision - Represents information about the version (or revision) of an action.

data PutActionRevision

Represents the input of a put action revision action.

See: putActionRevision smart constructor.

Request Lenses

pPipelineName :: Lens' PutActionRevision Text

The name of the pipeline that will start processing the revision to the source.

pStageName :: Lens' PutActionRevision Text

The name of the stage that contains the action that will act upon the revision.

pActionName :: Lens' PutActionRevision Text

The name of the action that will process the revision.

pActionRevision :: Lens' PutActionRevision ActionRevision

Represents information about the version (or revision) of an action.

Destructuring the Response

putActionRevisionResponse

Arguments

:: Int
prsResponseStatus

-> PutActionRevisionResponse

Creates a value of PutActionRevisionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prsNewRevision - Indicates whether the artifact revision was previously used in an execution of the specified pipeline.

prsPipelineExecutionId - The ID of the current workflow state of the pipeline.

prsResponseStatus - -- | The response status code.

data PutActionRevisionResponse

Represents the output of a put action revision action.

See: putActionRevisionResponse smart constructor.

Response Lenses

prsNewRevision :: Lens' PutActionRevisionResponse (Maybe Bool)

Indicates whether the artifact revision was previously used in an execution of the specified pipeline.

prsPipelineExecutionId :: Lens' PutActionRevisionResponse (Maybe Text)

The ID of the current workflow state of the pipeline.

prsResponseStatus :: Lens' PutActionRevisionResponse Int

- | The response status code.