Network.AWS.CodePipeline.PutJobFailureResult

Description

Represents the failure of a job as returned to the pipeline by a job worker. Only used for custom actions.

Creating a Request

putJobFailureResult

Arguments

:: Text
pjfrJobId

-> FailureDetails
pjfrFailureDetails

-> PutJobFailureResult

Creates a value of PutJobFailureResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pjfrJobId - The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

pjfrFailureDetails - The details about the failure of a job.

data PutJobFailureResult

Represents the input of a put job failure result action.

See: putJobFailureResult smart constructor.

Request Lenses

pjfrJobId :: Lens' PutJobFailureResult Text

The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

pjfrFailureDetails :: Lens' PutJobFailureResult FailureDetails

The details about the failure of a job.

Destructuring the Response

putJobFailureResultResponse :: PutJobFailureResultResponse

Creates a value of PutJobFailureResultResponse with the minimum fields required to make a request.

data PutJobFailureResultResponse

See: putJobFailureResultResponse smart constructor.