Network.AWS.CodePipeline.PutJobSuccessResult

Description

Represents the success of a job as returned to the pipeline by a job worker. Only used for custom actions.

Creating a Request

putJobSuccessResult

Arguments

:: Text
pjsrJobId

-> PutJobSuccessResult

Creates a value of PutJobSuccessResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pjsrContinuationToken - A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

pjsrExecutionDetails - The execution details of the successful job, such as the actions taken by the job worker.

pjsrCurrentRevision - The ID of the current revision of the artifact successfully worked upon by the job.

pjsrJobId - The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

data PutJobSuccessResult

Represents the input of a put job success result action.

See: putJobSuccessResult smart constructor.

Request Lenses

pjsrContinuationToken :: Lens' PutJobSuccessResult (Maybe Text)

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

pjsrExecutionDetails :: Lens' PutJobSuccessResult (Maybe ExecutionDetails)

The execution details of the successful job, such as the actions taken by the job worker.

pjsrCurrentRevision :: Lens' PutJobSuccessResult (Maybe CurrentRevision)

The ID of the current revision of the artifact successfully worked upon by the job.

pjsrJobId :: Lens' PutJobSuccessResult Text

The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

Destructuring the Response

putJobSuccessResultResponse :: PutJobSuccessResultResponse

Creates a value of PutJobSuccessResultResponse with the minimum fields required to make a request.

data PutJobSuccessResultResponse

See: putJobSuccessResultResponse smart constructor.