Network.AWS.CodePipeline.PutThirdPartyJobSuccessResult

Description

Represents the success of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

Creating a Request

putThirdPartyJobSuccessResult

Arguments

:: Text
ptpjsrJobId

-> Text
ptpjsrClientToken

-> PutThirdPartyJobSuccessResult

Creates a value of PutThirdPartyJobSuccessResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptpjsrContinuationToken - A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

ptpjsrExecutionDetails - The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

ptpjsrCurrentRevision - Represents information about a current revision.

ptpjsrJobId - The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

ptpjsrClientToken - The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

data PutThirdPartyJobSuccessResult

Represents the input of a put third party job success result action.

See: putThirdPartyJobSuccessResult smart constructor.

Request Lenses

ptpjsrContinuationToken :: Lens' PutThirdPartyJobSuccessResult (Maybe Text)

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

ptpjsrExecutionDetails :: Lens' PutThirdPartyJobSuccessResult (Maybe ExecutionDetails)

The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

ptpjsrCurrentRevision :: Lens' PutThirdPartyJobSuccessResult (Maybe CurrentRevision)

Represents information about a current revision.

ptpjsrJobId :: Lens' PutThirdPartyJobSuccessResult Text

The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

ptpjsrClientToken :: Lens' PutThirdPartyJobSuccessResult Text

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Destructuring the Response

putThirdPartyJobSuccessResultResponse :: PutThirdPartyJobSuccessResultResponse

Creates a value of PutThirdPartyJobSuccessResultResponse with the minimum fields required to make a request.

data PutThirdPartyJobSuccessResultResponse

See: putThirdPartyJobSuccessResultResponse smart constructor.