Network.AWS.CodePipeline.RetryStageExecution

Description

Resumes the pipeline execution by retrying the last failed actions in a stage.

Creating a Request

retryStageExecution

Arguments

:: Text
rsePipelineName

-> Text
rseStageName

-> Text
rsePipelineExecutionId

-> StageRetryMode
rseRetryMode

-> RetryStageExecution

Creates a value of RetryStageExecution with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsePipelineName - The name of the pipeline that contains the failed stage.

rseStageName - The name of the failed stage to be retried.

rsePipelineExecutionId - The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

rseRetryMode - The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

data RetryStageExecution

Represents the input of a retry stage execution action.

See: retryStageExecution smart constructor.

Request Lenses

rsePipelineName :: Lens' RetryStageExecution Text

The name of the pipeline that contains the failed stage.

rseStageName :: Lens' RetryStageExecution Text

The name of the failed stage to be retried.

rsePipelineExecutionId :: Lens' RetryStageExecution Text

The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

rseRetryMode :: Lens' RetryStageExecution StageRetryMode

The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

Destructuring the Response

retryStageExecutionResponse

Arguments

:: Int
rsersResponseStatus

-> RetryStageExecutionResponse

Creates a value of RetryStageExecutionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsersPipelineExecutionId - The ID of the current workflow execution in the failed stage.

rsersResponseStatus - -- | The response status code.

data RetryStageExecutionResponse

Represents the output of a retry stage execution action.

See: retryStageExecutionResponse smart constructor.

Response Lenses

rsersPipelineExecutionId :: Lens' RetryStageExecutionResponse (Maybe Text)

The ID of the current workflow execution in the failed stage.

rsersResponseStatus :: Lens' RetryStageExecutionResponse Int

- | The response status code.