Network.AWS.CodePipeline.UpdatePipeline

Description

Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure in conjunction with UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.

Creating a Request

updatePipeline

Arguments

:: PipelineDeclaration
upPipeline

-> UpdatePipeline

Creates a value of UpdatePipeline with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upPipeline - The name of the pipeline to be updated.

data UpdatePipeline

Represents the input of an update pipeline action.

See: updatePipeline smart constructor.

Request Lenses

upPipeline :: Lens' UpdatePipeline PipelineDeclaration

The name of the pipeline to be updated.

Destructuring the Response

updatePipelineResponse

Arguments

:: Int
uprsResponseStatus

-> UpdatePipelineResponse

Creates a value of UpdatePipelineResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uprsPipeline - The structure of the updated pipeline.

uprsResponseStatus - -- | The response status code.

data UpdatePipelineResponse

Represents the output of an update pipeline action.

See: updatePipelineResponse smart constructor.

Response Lenses

uprsPipeline :: Lens' UpdatePipelineResponse (Maybe PipelineDeclaration)

The structure of the updated pipeline.

uprsResponseStatus :: Lens' UpdatePipelineResponse Int

- | The response status code.